Relocate XML extensions and use empty sequences
