+++ src/MooVC.Syntax.CSharp/BooleanExtensions.ToXmlAttribute.cs
@@ -1,7 +1,7 @@
 namespace MooVC.Syntax.CSharp
 {
-    using System;
     using System.Collections.Generic;
+    using System.Linq;
     using System.Xml.Linq;
 
     internal static class BooleanExtensions
@@ -10,7 +10,7 @@ namespace MooVC.Syntax.CSharp
         {
             if (!value)
             {
-                return Array.Empty<object>();
+                return Enumerable.Empty<object>();
             }
 
             return new object[] { new XAttribute(name, value.ToString().ToLowerInvariant()) };
