+++ src/MooVC.Syntax.CSharp/SnippetExtensions.ToXmlAttribute.cs
@@ -1,7 +1,7 @@
 namespace MooVC.Syntax.CSharp
 {
-    using System;
     using System.Collections.Generic;
+    using System.Linq;
     using System.Xml.Linq;
 
     internal static class SnippetExtensions
@@ -10,7 +10,7 @@ namespace MooVC.Syntax.CSharp
         {
             if (value.IsEmpty)
             {
-                return Array.Empty<object>();
+                return Enumerable.Empty<object>();
             }
 
             return new object[] { new XAttribute(name, value.ToString()) };
